requirements.txt
gradio==4.44.1
pydub
python-dotenv
groq
